# swagger_ui/retrieve/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Definition of ArmRay backend API for data retrieval operations"
  version: "0.0.1"
  title: "ArmRay API"
host: "ec2-52-36-38-196.us-west-2.compute.amazonaws.com:1151"
basePath: "/armray/api/0.1/rest/retrieve"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /Skill/{id}:
    get:
      tags:
      - "Skill"
      description: "Retrieves `Skill` objects by **id**.\n"
      operationId: "getSkillById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Skill`"
          schema:
            $ref: "#/definitions/Skill"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
  /Skill/{id}/created:
    get:
      tags:
      - "Skill"
      description: "Gets `Skill` created date\n"
      operationId: "getSkillCreatedDateById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Skill` create date"
          schema:
            type: "string"
            format: "date-time"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
  /Skill/{id}/description:
    get:
      tags:
      - "Skill"
      description: "Gets description of `Skill`\n"
      operationId: "getSkillBodyById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Skill` body"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
  /Skill/{id}/description/{description}:
    get:
      tags:
      - "Skill"
      description: "Gets if description of `Skill` is on `Skill`\n"
      operationId: "getIsSkillBodyById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      - name: "description"
        in: "path"
        description: "The `Skill` description"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
  /Skill/{id}/updated:
    get:
      tags:
      - "Skill"
      description: "Gets `Skill` updated date\n"
      operationId: "getSkillUpdatedDateById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Skill` update date"
          schema:
            type: "string"
            format: "date-time"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
...
  /User/{id}/skills/{skill_id}:
    get:
      tags:
      - "User"
      - "Skill"
      description: "Check if `Skill`is associated with a `User`.\n"
      operationId: "getIsSkillOnUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "skill_id"
        in: "path"
        description: "The `id` of the `Skill` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of UserSkills ids"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
definitions:
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      fields:
        type: "string"
      message:
        type: "string"
...
  Skill:
    type: "object"
    properties:
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Skill`creation"
      description:
        type: "string"
        description: "Description of `Skill`"
      review_id:
        type: "string"
        description: "id of skill"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Skill`update"
...